fix: only memory address apply the uint64_t, the indices of memories still are uint32_t

+++ include/runtime/instance/module.h
@@ -416,7 +416,7 @@ protected:
     }
     return unsafeGetMemory(Idx);
   }
-  MemoryInstance *unsafeGetMemory(uint64_t Idx) const noexcept {
+  MemoryInstance *unsafeGetMemory(uint32_t Idx) const noexcept {
     return MemInsts[Idx];
   }
   TagInstance *unsafeGetTag(uint32_t Idx) const noexcept {
@@ -452,7 +452,7 @@ protected:
     }
     return unsafeGetData(Idx);
   }
-  DataInstance *unsafeGetData(uint64_t Idx) const noexcept {
+  DataInstance *unsafeGetData(uint32_t Idx) const noexcept {
     return DataInsts[Idx];
   }
